requirements.txt
duckdb==1.4.3
numpy==2.4.1
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.3

# Streaming
kafka-python-ng==2.2.2
